+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.8.0 - 23/01/2020
+
+- Added `onTap` property to `TextFieldConfiguration`
+- Added `offsetX` property to `SuggestionsBoxDecoration` and `CupertinoSuggestionsBoxDecoration`
+- Bug fixes
+
 ## 1.7.0 - 16/10/2019
 
 - Updated keyboard_visibility dependency